api_gateway/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
PyYAML
pytest
